scanners/tls-scanner/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asyncio-nats-client==0.11.4
cffi==1.15.0
cryptography==3.4.8; python_version >= '3.6'
emoji==1.6.1
httptools==0.3.0
nassl==4.0.0; python_version >= '3.7'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pynacl==1.4.0
pyopenssl==21.0.0
python-dotenv==0.19.1
scapy==2.4.5
setproctitle==1.2.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sslyze==5.0.3
